Moving header configuration out of main source code folder.
This makes embedding Sharekit as a github submodule easier, since you
can add the submodule, then in xcode only include the ShareKit
directory, and outside of the submodule directory, in your main git
repo, provide a custom SHKConfig.h.

The idea is not to have to overwrite the SHKConfig.h file provided by
the submodule and prevent changes to it propagating outside of your
repo. It makes installation/usage a little bit more complex, though.